Fixed ambiguous code and game loading
